==> src/02-bremen_lidoxml_to_csv.ipynb <==
# Convert LIDO XML to CSV format (Bremen repository)

### Create a csv file for all data ###

import os
import csv
import pandas as pd
import numpy as np
from lxml import etree
import requests

# Directory containing the XML files
folder_path = "../data/bremen/lido"
namespace = {"lido": "http://www.lido-schema.org"}

# Output CSV file path
output_csv = "../output_csv/bremen_output_with_placeOfBirth.csv"

# Initialize a list to store all record data
data = []
artist_info_dict = {}

# Loop through each subfolder in the lido directory
for subfolder in os.listdir(folder_path):
    subfolder_path = os.path.join(folder_path, subfolder)
    if os.path.isdir(subfolder_path):
        # Loop through each XML file in the subfolder
        for filename in os.listdir(subfolder_path):
            if filename.endswith(".xml"):
                filepath = os.path.join(subfolder_path, filename)
                tree = etree.parse(filepath)

                # Fields
                lido_record_id = tree.xpath("string(//lido:lidoRecID)", namespaces=namespace)
                image_urls = tree.xpath("//lido:linkResource/text()", namespaces=namespace)
                types = tree.xpath("//lido:objectWorkType/lido:term/text()", namespaces=namespace)
                materials = tree.xpath("//lido:termMaterialsTech/lido:term/text()", namespaces=namespace)
                display_date = tree.xpath("string(//lido:eventDate/lido:displayDate)", namespaces=namespace)
                earliest_date = tree.xpath("string(//lido:eventDate/lido:date/lido:earliestDate)", namespaces=namespace)
                latest_date = tree.xpath("string(//lido:eventDate/lido:date/lido:latestDate)", namespaces=namespace)
                subjects = tree.xpath("//lido:subjectConcept/lido:term/text()", namespaces=namespace)
                artist_name = tree.xpath("string(//lido:nameActorSet/lido:appellationValue[@lido:pref='preferred'])", namespaces=namespace)
                actor_gnd_id = tree.xpath("string(//lido:actorID[@lido:pref='preferred'])", namespaces=namespace)
                gender = None
                placeOfBirth = None
                placeOfDeath = None
                placeOfActivity = None
                associatedCountry = None
                role = tree.xpath("string(//lido:roleActor/lido:term)", namespaces=namespace)
                birth = tree.xpath("string(//lido:vitalDatesActor/lido:earliestDate)", namespaces=namespace)
                death = tree.xpath("string(//lido:vitalDatesActor/lido:latestDate)", namespaces=namespace)
                location = tree.xpath("string(//lido:repositoryLocation/lido:namePlaceSet/lido:appellationValue)", namespaces=namespace)
                classifications = tree.xpath("//lido:classification/lido:term/text()", namespaces=namespace)
                titles = tree.xpath("//lido:titleSet/lido:appellationValue/text()", namespaces=namespace)
                rights_statements = tree.xpath("//lido:rightsResource/lido:rightsType/lido:conceptID[@lido:type='http://terminology.lido-schema.org/lido00099']/text()", namespaces=namespace)
                work_id = tree.xpath("string(//lido:workID)", namespaces=namespace)
                repository_name = tree.xpath("//lido:legalBodyName/lido:appellationValue/text()", namespaces=namespace)
                legal_body_link = tree.xpath("string(//lido:legalBodyWeblink)", namespaces=namespace)
                record_id = tree.xpath("string(//lido:recordID)", namespaces=namespace)
                record_links = tree.xpath("//lido:recordInfoLink/text()", namespaces=namespace)
                record_metadata_date = tree.xpath("string(//lido:recordMetadataDate)", namespaces=namespace)
                event_type = tree.xpath("string(//lido:eventType/lido:term)", namespaces=namespace)
                display_materials = tree.xpath("//lido:displayMaterialsTech/text()", namespaces=namespace)
                inscriptions = tree.xpath("//lido:inscriptionDescription/lido:descriptiveNoteValue/text()", namespaces=namespace)
                related_work_notes = tree.xpath("//lido:objectNote/text()", namespaces=namespace)
                resource_types = tree.xpath("//lido:resourceType/lido:term/text()", namespaces=namespace)
                credit_line = tree.xpath("string(//lido:creditLine)", namespaces=namespace)
                object_description = tree.xpath("string(//lido:objectDescriptionSet/lido:descriptiveNoteValue)", namespaces=namespace)
                display_measurements = tree.xpath("string(//lido:displayObjectMeasurements)", namespaces=namespace)


                # Get gender and nationality of the artist from entity facts with his/her gnd id
                if actor_gnd_id != "":
                    actor_gnd_id = actor_gnd_id.split("/")[-1]

                    if (actor_gnd_id in artist_info_dict):
                        gender = artist_info_dict[actor_gnd_id]["gender"]
                        placeOfBirth = artist_info_dict[actor_gnd_id]["placeOfBirth"]
                        placeOfDeath = artist_info_dict[actor_gnd_id]["placeOfDeath"]
                        placeOfActivity = artist_info_dict[actor_gnd_id]["placeOfActivity"]
                        associatedCountry = artist_info_dict[actor_gnd_id]["associatedCountry"]
                    else:
                        url = f"https://hub.culturegraph.org/entityfacts/{actor_gnd_id}"
                        response = requests.get(url)

                        if response.status_code == 200:
                            gnd_data = response.json()
                            gender = gnd_data.get("gender", {}).get("label", "")
                            placeOfBirth = "; ".join(
                                [entry.get("preferredName", "") for entry in gnd_data.get("placeOfBirth", [])]
                            ) or ""

                            placeOfDeath = "; ".join(
                                [entry.get("preferredName", "") for entry in gnd_data.get("placeOfDeath", [])]
                            ) or ""

                            placeOfActivity = "; ".join(
                                [entry.get("preferredName", "") for entry in gnd_data.get("placeOfActivity", [])]
                            ) or ""

                            associatedCountry_list = gnd_data.get("associatedCountry", [])
                            associatedCountry = "; ".join([n.get("preferredName", "") for n in associatedCountry_list]) if associatedCountry_list else ""
                            
                            artist_info_dict[actor_gnd_id] = {"gender": gender, "placeOfBirth": placeOfBirth, "placeOfDeath": placeOfDeath, "placeOfActivity": placeOfActivity, "associatedCountry": associatedCountry}
                        else:
                            print(f"Failed to fetch GND data: {response.status_code} for {url}")

                data.append({
                    "lidoRecordId": lido_record_id,
                    "imageUrl": "; ".join(image_urls),
                    "type": "; ".join(types),
                    "material": "; ".join(materials),
                    "displayDate": display_date,
                    "earliestDate": earliest_date,
                    "latestDate": latest_date,
                    "subject": "; ".join(subjects),
                    "artistName (preferred)": artist_name,
                    "artistGNDId": actor_gnd_id,
                    "genderOfArtist": gender,
                    "placeOfBirth": placeOfBirth,
                    "placeOfDeath": placeOfDeath,
                    "placeOfActivity": placeOfActivity,
                    "associatedCountry": associatedCountry,
                    "role": role,
                    "birth year": birth,
                    "death year": death,
                    "location": location,
                    "title": "; ".join(titles),
                    "classification": "; ".join(classifications),
                    "rightsStatement": "; ".join(rights_statements),
                    "workID": work_id,
                    "repositoryName": repository_name,
                    "recordID": record_id,
                    "recordLinks": "; ".join(record_links),
                    "recordMetadataDate": record_metadata_date,
                    "eventType": event_type,
                    "displayMaterialsTech": "; ".join(display_materials),
                    "displayMeasurements": display_measurements,
                    "objectDescription": object_description,
                    "inscriptions": "; ".join(inscriptions),
                    "relatedWorkNotes": "; ".join(related_work_notes),
                    "resourceType": "; ".join(resource_types),
                    "creditLine": credit_line
                })

# Convert to DataFrame and save to CSV
df = pd.DataFrame(data)
df.replace("", np.nan, inplace=True)
df.to_csv(output_csv, index=False, encoding="utf-8")

import pandas as pd

# Output CSV file path
output_csv = "../output_csv/bremen_output_with_placeOfBirth.csv"

# Load and display the CSV as a table
df_bremen = pd.read_csv(output_csv)
print(f"The full dataset has {df_bremen.shape[0]} rows and {df_bremen.shape[1]} columns.")
print("First ten rows of the dataset")
df_bremen.head(10)
