==> src/03-bremen_data_profiling.ipynb <==
<h3>Use ydata-profiling to generate report on the data</h3>

# Make sure to have ydata_profiling package installed, python version 3.13.0 is not supported
#!pip install ydata-profiling
# Probably also need to run if you get a ModuleNotFoundError
# !pip install --upgrade setuptools

import pandas as pd
from ydata_profiling import ProfileReport

df = pd.read_csv("../output_csv/bremen_output.csv")

profile = ProfileReport(df, title="Bremen Data Profiling Report")
profile.to_file("../output_csv/bremen_data_profiling_report.html")
