==> src/07-artist_placeOfBirth_map.ipynb <==
### Create Artist's Birth Place Map over time ###

# import of the libraries
# !pip install mapclassify
# !pip install folium

import pandas as pd
import re

artist_csv = "../output_csv/bremen_data_v5_merge_artist.csv"

df_artist = pd.read_csv(artist_csv)
df_artist.head(10)

total_num = df_artist.shape[0]
count_anonym = df_artist[df_artist['artistName (preferred)'].str.contains("Anonym", na=False)].shape[0]

print(f"Total number of artists in Bremen dataset: {total_num}")
print(f"Number of anonymous artists: {count_anonym} which is {count_anonym/total_num*100}%")

(df_artist["placeOfBirth"].isna() | (df_artist["placeOfBirth"] == "")).sum()

### Convert birth place to longitude and latitude ###

# pip install pandas geopy
# !pip install shapely

from geopy.geocoders import Nominatim
from geopy.extra.rate_limiter import RateLimiter
from shapely.geometry import Point
import pandas as pd
import numpy as np

# Set up geocoder
geolocator = Nominatim(user_agent="geoapi")
geocode = RateLimiter(geolocator.geocode, min_delay_seconds=1)

# Prepare latitude and longitude columns


len(df_artist["placeOfBirth"].unique())

city_locations_dict = {}

# Loop through each place and get coordinates
for place in df_artist["placeOfBirth"].unique():
    try:
        location = geocode(place)

        if location:
            city_locations_dict[place] = {"latitude": location.latitude, "longitude": location.longitude}
    except Exception as e:
        print(f"Error with place '{place}': {e}")

print(city_locations_dict)

df_artist["latitude"] = df_artist["placeOfBirth"].apply(lambda x: city_locations_dict.get(x, {}).get("latitude"))
df_artist["longitude"] = df_artist["placeOfBirth"].apply(lambda x: city_locations_dict.get(x, {}).get("longitude"))

df_artist.head(10)

output_csv = "../output_csv/bremen_data_v5_2_with_coordinate.csv"

df_artist.to_csv(output_csv, index=False, encoding="utf-8")

import folium
import pandas as pd

coordinative_csv = "../output_csv/bremen_data_v5_2_with_coordinate.csv"
df_artist_with_coordinate = pd.read_csv(coordinative_csv) 
df_artist_with_coordinate["birth year"] = df_artist_with_coordinate["birth year"].astype("Int64")
df_artist_with_coordinate["birth_year_span_start"] = df_artist_with_coordinate["birth_year_span_start"].astype("Int64")

df_artist_with_coordinate.head()


df_artist_with_coordinate["placeOfBirth"].value_counts()


df_artist_with_coordinate["birth_year_span_start"].value_counts()

Split dataframe to multiple dataframes based on `birth_year_span_start`

grouped_dfs = {
    year: group for year, group in df_artist_with_coordinate.groupby("birth_year_span_start")
}
print(grouped_dfs.keys())
grouped_dfs[1400].head()

merged_dfs_by_year = {}

for year, group in grouped_dfs.items():
    merged_df = group.groupby("placeOfBirth").agg({
        "artistName (preferred)": lambda x: "; ".join(x.dropna().unique()),
        "placeOfBirth": "count",  # count how many artists per place
        "latitude": "first",
        "longitude": "first"
    }).rename(columns={
        "artistName (preferred)": "artists",
        "placeOfBirth": "artistCount"
    }).reset_index()
    
    merged_dfs_by_year[year] = merged_df

merged_dfs_by_year[1400].head()

# assigns a color on the map marker based on artistCount
def get_color(count):
    if count >= 15:
        return 'darkred'
    elif count >= 10:
        return 'red'
    elif count >= 5:
        return 'orange'
    elif count >= 3:
        return 'blue'
    else:
        return 'green'

for year, df in merged_dfs_by_year.items():
    m = folium.Map(location=[50, 10], zoom_start=4)

    for _, row in df.dropna(subset=["latitude", "longitude"]).iterrows():
        folium.CircleMarker(
            location=[row["latitude"], row["longitude"]],
            radius=6,
            popup=folium.Popup(f"<b>Artists:</b> {row['artists']}<br><b>Count:</b> {row['artistCount']}", max_width=300),
            color=get_color(row["artistCount"]),
            fill=True,
            fill_color=get_color(row["artistCount"]),
            fill_opacity=0.8
        ).add_to(m)

    # Save the map for this year
    m.save(f"../output_map/artist_map_{year}.html")
