==> src/07_1_map.ipynb <==
import folium
import pandas as pd

coordinative_csv = "../output_csv/bremen_data_v5_2_with_coordinate.csv"
df_artist_with_coordinate = pd.read_csv(coordinative_csv) 
df_artist_with_coordinate["birth year"] = df_artist_with_coordinate["birth year"].astype("Int64")
df_artist_with_coordinate["birth_year_span_start"] = df_artist_with_coordinate["birth_year_span_start"].astype("Int64")

df_artist_with_coordinate.head()


df_artist_with_coordinate["placeOfBirth"].value_counts()


df_artist_with_coordinate["birth_year_span_start"].value_counts()

Split dataframe to multiple dataframes based on `birth_year_span_start`

grouped_dfs = {
    year: group for year, group in df_artist_with_coordinate.groupby("birth_year_span_start")
}
print(grouped_dfs.keys())
grouped_dfs[1400].head()

merged_dfs_by_year = {}

for year, group in grouped_dfs.items():
    merged_df = group.groupby("placeOfBirth").agg({
        "artistName (preferred)": lambda x: "; ".join(x.dropna().unique()),
        "placeOfBirth": "count",  # count how many artists per place
        "latitude": "first",
        "longitude": "first"
    }).rename(columns={
        "artistName (preferred)": "artists",
        "placeOfBirth": "artistCount"
    }).reset_index()
    
    merged_dfs_by_year[year] = merged_df

merged_dfs_by_year[1400].head()

import ipywidgets as widgets
from IPython.display import display, clear_output
from branca.element import Template, MacroElement

year_slider = widgets.IntSlider(
    value=min(merged_dfs_by_year.keys()),
    min=1450,
    max=1900,
    step=50, 
    description='Year:'
)

def update_map(year):
    clear_output(wait=True)
    display(year_slider)
    
    df = merged_dfs_by_year[year]

    # Create a new map
    m = folium.Map(location=[50, 10], zoom_start=4)

    # Color based on artist count
    for _, row in df.dropna(subset=["latitude", "longitude"]).iterrows():
        count = row["artistCount"]
        if count > 10:
            color = "darkred"
        elif count >= 5:
            color = "orange"
        elif count >=  3:
            color = "blue"
        else:
            color = "green"

        folium.CircleMarker(
            location=[row["latitude"], row["longitude"]],
            # radius=5 + count * 0.3,
            radius= 6,
            popup=folium.Popup(f"<b>Place of Birth:</b> {row['placeOfBirth']}<br> <b>Artists:</b> {row['artists']}<br><b>Count:</b> {row['artistCount']}", max_width=300),
            color=color,
            fill=True,
            fill_opacity=0.7
        ).add_to(m)
    
    legend_html = """
    {% macro html(this, kwargs) %}
    <div style="
        position: fixed;
        bottom: 50px;
        left: 50px;
        width: 150px;
        height: 120px;
        background-color: white;
        border:2px solid grey;
        z-index:9999;
        font-size:14px;
        padding: 10px;
        ">
        <b>Artist Count</b><br>
        <i style="color:darkred;">●</i> > 10<br>
        <i style="color:orange;">●</i> 5–10<br>
        <i style="color:blue;">●</i> 3–4<br>
        <i style="color:green;">●</i> 1–2<br>
    </div>
    {% endmacro %}
    """
    legend = MacroElement()
    legend._template = Template(legend_html)
    m.get_root().add_child(legend)

    display(m)

# 3. Bind slider to map update
year_slider.observe(lambda change: update_map(change['new']), names='value')

# 4. Display the slider and the first map
display(year_slider)
update_map(year_slider.value)
